+++ ts/gulpfile.ts
@@ -1,22 +1,21 @@
+import gulp from "gulp";
+import minimist from "minimist";
 import * as buildTools from "turbo-gulp";
+import { Project } from "turbo-gulp/project";
 import { LibTarget, registerLibTasks } from "turbo-gulp/targets/lib";
 import { MochaTarget, registerMochaTasks } from "turbo-gulp/targets/mocha";
 import { NodeTarget, registerNodeTasks } from "turbo-gulp/targets/node";
 
-import gulp from "gulp";
-import minimist from "minimist";
-
 interface Options {
-  devDist?: string;
+  next?: string;
 }
 
 const options: Options & minimist.ParsedArgs = minimist(process.argv.slice(2), {
-  string: ["devDist"],
-  default: {devDist: undefined},
-  alias: {devDist: "dev-dist"},
+  string: ["next"],
+  default: {next: undefined},
 });
 
-const project: buildTools.Project = {
+const project: Project = {
   root: __dirname,
   packageJson: "package.json",
   buildDir: "build",
@@ -25,12 +24,10 @@ const project: buildTools.Project = {
   tslint: {
     configuration: {
       rules: {
-        "max-file-line-count": false,
-        "no-submodule-imports": false,
         "whitespace": false,
-      },
-    },
-  },
+      }
+    }
+  }
 };
 
 const lib: LibTarget = {
@@ -41,7 +38,7 @@ const lib: LibTarget = {
   mainModule: "index",
   dist: {
     packageJsonMap: (old: buildTools.PackageJson): buildTools.PackageJson => {
-      const version: string = options.devDist !== undefined ? `${old.version}-build.${options.devDist}` : old.version;
+      const version: string = options.next !== undefined ? `${old.version}-build.${options.next}` : old.version;
       return <any> {...old, version, scripts: undefined, private: false};
     },
     npmPublish: {
@@ -61,11 +58,6 @@ const lib: LibTarget = {
       branch: "gh-pages",
     },
   },
-  copy: [
-    {
-      files: ["**/*.json"],
-    },
-  ],
   clean: {
     dirs: ["build/lib", "dist/lib"],
   },
@@ -80,13 +72,6 @@ const test: MochaTarget = {
   tscOptions: {
     skipLibCheck: true,
   },
-  copy: [
-    {
-      src: "../../test",
-      files: ["parsers/**/*.json"],
-      dest: "test",
-    },
-  ],
   clean: {
     dirs: ["build/test"],
   },