+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+
+### Typescript
+
+- **[Internal]** Update build tools.
+
 # 0.5.3 (2019-05-05)
 
 - **[Feature]** Implement parser for `DefineBinaryData` (thanks [@dmarcuse](https://github.com/dmarcuse)).